+++ cursus/Libft/ft_memcpy.c
@@ -10,9 +10,11 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memcpy(void *dest, const void *src, long unsigned int n)
+#include "libft.h"
+
+void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	long unsigned int	i;
+	size_t	i;
 	char				*cast_dest;
 	const char			*cast_src;
