+++ cursus/Libft/no_ft_substr.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-char	*ft_substr(char const *s, unsigned int start, long unsigned int len)
+char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
 	
 }
